requirements.txt
prompt-toolkit==3.0.5
pyyaml==5.3.1
wcwidth==0.2.5
